fix(config): Make ONNX Runtime DirectML an optional configuration for better compatibility
This commit introduces an optional configuration for using ONNX Runtime DirectML.

DirectML can significantly improve performance on newer devices, but it has stricter requirements and may not be suitable for all users.

By making it an optional configuration, we ensure better compatibility for users with older devices or who are running the CLI version, where the GPU's potential may not be fully utilized due to single inference models.

For more information on DirectML requirements, please refer to: https://github.com/microsoft/DirectML